refactor(*): moved files, added barrel import, now index.ts only exposes bare minimum
