+++ packages/workflow-core/src/lib/create-workflow.ts
@@ -1,0 +1,14 @@
+import {TCreateWorkflow} from "./types";
+import {WorkflowRunner} from "./workflow-runner";
+
+export const createWorkflow: TCreateWorkflow = ({
+                                                  workflowDefinition,
+                                                  workflowActions,
+                                                  extensions,
+                                                }) =>
+  new WorkflowRunner({
+    workflowDefinition,
+    workflowActions,
+    context: {},
+    extensions,
+  });
