+++ packages/workflow-core/src/lib/index.ts
@@ -1,0 +1,8 @@
+export {
+  WorkflowEvent,
+  WorkflowOptions,
+  WorkflowRunnerArgs,
+  WorkflowEventWithoutState,
+  StatePlugin,
+} from "./types";
+export {createWorkflow} from "./create-workflow";
